🚧 PCQK99 release: split recovery report rendering
